backend/requirements.txt
fastapi[standard]
uvicorn[standard]
psycopg[binary]
sqlmodel
python-jose
bcrypt